arborescence: move processors in dedicated folders
